langgraph-pattern-practice/pyproject.toml
[project]
name = "langgraph-pattern-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
force-single-line = true
